fix(fleet): address prix review — extend to check.ts, gate fail-fast on fresh stats
- Move fleetDialTarget into lib/devices/connect.ts (next to sshTargetFor) so it's
  shared, not command-local; tests move to connect.test.ts.
- Extend the fix to agents check --devices (check.ts probeDeviceCheck), which had
  the identical bare-name bug prix flagged — now dials the registry address too.
  Verified live: 'check --devices' reaches all 12 (drift 12 of 12) where m1-m6
  previously errored on the stale ~/.ssh/config IP.
- Gate the fail-fast skip on forceRefresh (--refresh/--live): loadFleetStats is
  cache-first, so on a default run a box that came back online since the last
  daemon warm would be skipped to 'unreachable' without a live try — a false
  negative prix flagged. Now default runs always do the (now-fast) live probe.
